Version check

+++ src/os_win32.c
@@ -7690,6 +7690,12 @@ mch_setenv(char *var, char *value, int x)
  */
 #define VTP_FIRST_SUPPORT_BUILD MAKE_VER(10, 0, 15063)
 
+/*
+ * Support for pseudo-console (ConPTY) was added in windows 10
+ * version 1809 (October 2018 update).
+ */
+#define CONPTY_FIRST_SUPPORT_BUILD MAKE_VER(10, 0, 17763)
+
 #ifdef FEAT_GUI_W32
 static HWINSTA origsta;
 static HWINSTA newsta;
@@ -7701,7 +7707,8 @@ static PROCESS_INFORMATION pista;
     static void
 vtp_flag_init(void)
 {
-    DWORD   ver, mode;
+    DWORD   ver = get_build_number();
+    DWORD   mode;
     HANDLE  out;
 
 #ifdef FEAT_GUI_W32
@@ -7712,52 +7719,56 @@ vtp_flag_init(void)
     SECURITY_ATTRIBUTES sa;
     STARTUPINFO		si;
 
-    /* A console opened on another window station, get its standard output. */
-
-    vim_memset(&sa, 0, sizeof(sa));
-    sa.nLength = sizeof(SECURITY_ATTRIBUTES);
-    sa.bInheritHandle = TRUE;
-
-    origdesk = GetThreadDesktop(GetCurrentThreadId());
-
-    origsta = GetProcessWindowStation();
-
-    newsta = CreateWindowStation(NULL, 0, WINSTA_ALL_ACCESS, &sa);
-
-    if (SetProcessWindowStation(newsta))
+    if (ver >= CONPTY_FIRST_SUPPORT_BUILD)
     {
-	newdesk = CreateDesktop("Default", NULL, NULL, 0, GENERIC_ALL, &sa);
-	SetThreadDesktop(newdesk);
-	name[0] = NUL;
-	GetUserObjectInformation(newsta, UOI_NAME, name, sizeof(name), NULL);
-	sprintf(startup, "%s\\Default", name);
+	/* A console opened on another window station, get its standard
+	 * output. */
 
-	vim_memset(&si, 0, sizeof(si));
-	si.cb = sizeof(STARTUPINFO);
-	si.lpDesktop = startup;
+	vim_memset(&sa, 0, sizeof(sa));
+	sa.nLength = sizeof(SECURITY_ATTRIBUTES);
+	sa.bInheritHandle = TRUE;
 
-        vim_memset(&pista, 0, sizeof(pista));
+	origdesk = GetThreadDesktop(GetCurrentThreadId());
+	origsta = GetProcessWindowStation();
 
-	cmd = mch_getenv("COMSPEC");
-	if (cmd == NULL || *cmd == NUL)
-	    cmd = (char_u *)default_shell();
-	sprintf(cmdline, "%s /k", cmd);
+	newsta = CreateWindowStation(NULL, 0, WINSTA_ALL_ACCESS, &sa);
 
-	CreateProcess((char *)cmd, cmdline, NULL, NULL, TRUE,
+	if (SetProcessWindowStation(newsta))
+	{
+	    newdesk = CreateDesktop("Default", NULL, NULL, 0, GENERIC_ALL,
+									  &sa);
+	    SetThreadDesktop(newdesk);
+	    name[0] = NUL;
+	    GetUserObjectInformation(newsta, UOI_NAME, name, sizeof(name),
+									 NULL);
+	    sprintf(startup, "%s\\Default", name);
+
+	    vim_memset(&si, 0, sizeof(si));
+	    si.cb = sizeof(STARTUPINFO);
+	    si.lpDesktop = startup;
+
+	    vim_memset(&pista, 0, sizeof(pista));
+
+	    cmd = mch_getenv("COMSPEC");
+	    if (cmd == NULL || *cmd == NUL)
+		cmd = (char_u *)default_shell();
+	    sprintf(cmdline, "%s /k", cmd);
+
+	    CreateProcess((char *)cmd, cmdline, NULL, NULL, TRUE,
 				  CREATE_NEW_CONSOLE, NULL, NULL, &si, &pista);
-	Sleep(100);
-	AttachConsole(pista.dwProcessId);
+	    Sleep(100);
+	    AttachConsole(pista.dwProcessId);
 
-	freopen("CONOUT$", "w", stdout);
-	SetStdHandle(STD_OUTPUT_HANDLE, CreateFile("CONOUT$",
-		GENERIC_READ | GENERIC_WRITE,
-		FILE_SHARE_READ | FILE_SHARE_WRITE,
-		&sa, OPEN_EXISTING, 0, NULL));
+	    freopen("CONOUT$", "w", stdout);
+	    SetStdHandle(STD_OUTPUT_HANDLE, CreateFile("CONOUT$",
+		    GENERIC_READ | GENERIC_WRITE,
+		    FILE_SHARE_READ | FILE_SHARE_WRITE,
+		    &sa, OPEN_EXISTING, 0, NULL));
+	}
     }
 #endif
     out = GetStdHandle(STD_OUTPUT_HANDLE);
 
-    ver = get_build_number();
     vtp_working = (ver >= VTP_FIRST_SUPPORT_BUILD) ? 1 : 0;
     GetConsoleMode(out, &mode);
     mode |= (ENABLE_PROCESSED_OUTPUT | ENABLE_VIRTUAL_TERMINAL_PROCESSING);
@@ -7765,8 +7776,11 @@ vtp_flag_init(void)
 	vtp_working = 0;
 
 #ifdef FEAT_GUI_W32
-    SetThreadDesktop(origdesk);
-    SetProcessWindowStation(origsta);
+    if (ver >= CONPTY_FIRST_SUPPORT_BUILD)
+    {
+	SetThreadDesktop(origdesk);
+	SetProcessWindowStation(origsta);
+    }
 #endif
 }
 
@@ -7774,15 +7788,21 @@ vtp_flag_init(void)
 vtp_flag_exit(void)
 {
 #ifdef FEAT_GUI_W32
-    SetProcessWindowStation(newsta);
-    SetThreadDesktop(newdesk);
-    TerminateProcess(pista.hProcess, 0);
-    TerminateThread(pista.hThread, 0);
-    FreeConsole();
-    SetThreadDesktop(origdesk);
-    SetProcessWindowStation(origsta);
-    CloseWindowStation(newsta);
-    CloseDesktop(newdesk);
+    DWORD ver;
+
+    ver = get_build_number();
+    if (ver >= CONPTY_FIRST_SUPPORT_BUILD)
+    {
+	SetProcessWindowStation(newsta);
+	SetThreadDesktop(newdesk);
+	TerminateProcess(pista.hProcess, 0);
+	TerminateThread(pista.hThread, 0);
+	FreeConsole();
+	SetThreadDesktop(origdesk);
+	SetProcessWindowStation(origsta);
+	CloseWindowStation(newsta);
+	CloseDesktop(newdesk);
+    }
 #endif
 }
